[SYCL] Remove dependency on event_impl from cg.hpp
This patch is a part of effort to decouple SYCL Runtime library
interface from its actual implementation. The goal is to improve
SYCL ABI/API compatibility between different versions of library.

This patch moves HostProfilingInfo into a separate export-friendly
header file. It also removes the event_impl header from unwanted
places.
